6389 Extract and rename to ToggleButtonGroup
